requirements.txt
selenium
requests
python-dotenv
flask
